extract filter to a helper method
